+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 
 ## Unreleased
 
+- changed: Restrict ChangeHero trading to whitelisted plugins
 - changed: Replace deprecated currency codes in `SwapCurrencyError` with requests
 
 ## 0.21.9 (2023-10-02)